+++ app/src/main/java/org/android/bbangzip/data/datasource/remote/dto/request/RequestCompleteTimerDto.kt
@@ -1,0 +1,12 @@
+package org.android.bbangzip.data.datasource.remote.dto.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestCompleteTimerDto(
+    @SerialName("targetDate")
+    val targetDate: String,
+    @SerialName("count")
+    val count: Int,
+)
