+++ app/src/main/java/org/android/bbangzip/data/datasource/remote/dto/response/ResponseDummyDto.kt
@@ -1,0 +1,16 @@
+package org.android.bbangzip.data.datasource.remote.dto.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import org.android.bbangzip.domain.model.Dummy
+
+@Serializable
+data class ResponseDummyDto(
+    @SerialName("dummy")
+    val dummyName: String,
+) {
+    fun toDummyEntity() =
+        Dummy(
+            dummyName = dummyName,
+        )
+}
