+++ app/src/main/java/org/android/bbangzip/data/datasource/remote/dto/response/ResponseTodayBreadDto.kt
@@ -1,0 +1,10 @@
+package org.android.bbangzip.data.datasource.remote.dto.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseTodayBreadDto(
+    @SerialName("todayBakedCount")
+    val todayBakedCount: Int,
+)
