+++ app/src/main/java/org/android/bbangzip/data/repository/local/UserRepositoryImpl.kt
@@ -2,7 +2,7 @@ package org.android.bbangzip.data.repository.local
 
 import kotlinx.coroutines.flow.Flow
 import org.android.bbangzip.UserPreferences
-import org.android.bbangzip.data.source.local.datasource.UserLocalDataSource
+import org.android.bbangzip.data.datasource.local.UserLocalDataSource
 import org.android.bbangzip.domain.repository.UserRepository
 import javax.inject.Inject
