+++ app/src/main/java/org/android/bbangzip/data/repository/remote/DummyRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package org.android.bbangzip.data.repository.remote
 
-import org.android.bbangzip.data.source.remote.datasource.DummyRemoteDataSource
+import org.android.bbangzip.data.datasource.remote.DummyRemoteDataSource
 import org.android.bbangzip.domain.model.Dummy
 import org.android.bbangzip.domain.repository.DummyRepository
 import javax.inject.Inject
