--- app/src/main/java/org/android/bbangzip/data/source/remote/datasource/TimerRemoteDataSource.kt
@@ -1,26 +1,0 @@
-package org.android.bbangzip.data.source.remote.datasource
-
-import org.android.bbangzip.data.source.remote.dto.request.RequestCompleteTimerDto
-import org.android.bbangzip.data.source.remote.dto.response.ResponseBreadListDto
-import org.android.bbangzip.data.source.remote.dto.response.ResponseCompleteTimerDto
-import org.android.bbangzip.data.source.remote.dto.response.ResponseTodayBreadDto
-import org.android.bbangzip.data.source.remote.service.TimerService
-import org.android.bbangzip.data.source.remote.util.base.BaseResponse
-import javax.inject.Inject
-
-class TimerRemoteDataSource
-    @Inject
-    constructor(
-        private val timerService: TimerService,
-    ) {
-        // 오늘 구운 빵 개수 조회
-        suspend fun getTodayBreadCount(): BaseResponse<ResponseTodayBreadDto?> = timerService.getTodayBreadCount()
-
-        // 빵 리스트 조회
-        suspend fun getBreadList(): BaseResponse<ResponseBreadListDto?> = timerService.getBreadList()
-
-        // 타이머 완료 처리
-        suspend fun postTimerCompleted(
-            requestCompleteTimerDto: RequestCompleteTimerDto,
-        ): BaseResponse<ResponseCompleteTimerDto?> = timerService.postTimerCompleted(requestCompleteTimerDto)
-    }
